Extend E2E Docker tests to verify all files
The former test verified only the cache, which failed when the output
changed, but it made it very difficult to see in the diff _why_ the
test had failed and _why_ the cache snapshot had changed.

This commit extends the test to verify the contents of each file to make
it easier to see what effect a change had on the actual SQL output.